requirements.txt
mysql-connector-python==9.0.0
PyQt5==5.15.11
